apply customHook
