+++ src/App.css
@@ -192,5 +192,5 @@ h3 {
 }
 
 .skill-item {
-  margin: 0 48px;
+  margin: 0 auto;
 }
